Splits out form mapping into own file.
